requirements.txt
Flask
Flask_login
Flask_sqlalchemy
Flask_bcrypt
Flask_login
Flask-Migrate
Flask_wtf
wtforms
wtforms.validators
opencv-python
numpy
torch 
torchvision 
Pillow 
efficientnet_pytorch
json
